Pre dev

+++ src/providers/services/schedule.service.ts
@@ -1,6 +1,7 @@
-import { forwardRef, Inject, Injectable, InternalServerErrorException } from '@nestjs/common';
+import { ConflictException, forwardRef, HttpStatus, Inject, Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import * as moment from "moment";
+import { scheduled } from 'rxjs';
 import { Appointment } from 'src/appointments/entities/appointment.entity';
 import { AppointmentService } from 'src/appointments/services/appointment.service';
 import { Service } from 'src/facilities/entities/services.entity';
@@ -57,7 +58,7 @@ export class ScheduleService {
         const doctor = await this.doctorService.findOne(createScheduleInput.doctorId)
         scheduleInstance.doctor = doctor
       }
-      const schedule =  await this.scheduleRepository.save(scheduleInstance);
+      const schedule = await this.scheduleRepository.save(scheduleInstance);
       if(createScheduleInput.servicesIds){
         const services = await this.servicesService.findByIds(createScheduleInput.servicesIds)
         const serviceScheduleInstance = await this.createScheduleService(services, schedule.id)
@@ -310,7 +311,21 @@ export class ScheduleService {
    */
   async removeSchedule({ id }: RemoveSchedule) {
     try {
-      await this.scheduleRepository.delete(id)
+      const schedule = await this.findOne(id)
+      if(!schedule){
+        throw new NotFoundException({
+          status: HttpStatus.NOT_FOUND,
+          error: 'Schedule not found',
+        });
+      }
+      const appointmentExist = await this.appointmentService.findAppointmentByProviderId({offset: 0, serviceId: '', id: schedule.doctorId, currentDate: ""}, schedule.startAt, schedule.endAt)
+      if(appointmentExist.length){
+        throw new ConflictException({
+          status: HttpStatus.CONFLICT,
+          error: 'Appointment already booked with this schedule, can not delete it.',
+        });
+      }
+     await this.scheduleRepository.delete(id)
     } catch (error) {
       throw new InternalServerErrorException(error);
     }
